compile function from dirs, not tied to a project decoupling functionality (towards ethpm)
